+++ NEWS.md
@@ -1,3 +1,8 @@
+# heatwaveR 0.3.4 (2018-10-19)
+
+* `ts2clm()` no longer calculates variance column by default
+* `make_whole()` has been deprecated
+
 # heatwaveR 0.3.4 (2018-10-17)
 
 * All major functions now produce results only up to the fourth decimal place